% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{ArrayIntervalMethylation}
\alias{ArrayIntervalMethylation}
\title{ArrayIntervalMethylation}
\usage{
ArrayIntervalMethylation(MM, IntervalGR, cpgGR)
}
\description{
ArrayIntervalMethylation
}
\author{
Jiahao Wang
}
