% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{barplot2}
\alias{barplot2}
\title{barplot2}
\usage{
barplot2(
  data,
  outName = NULL,
  PDF = FALSE,
  PNG = TRUE,
  w = 6,
  h = 6,
  n = 20,
  title = "",
  padjust = TRUE,
  addLine = FALSE,
  fill = "#0089CB"
)
}
\description{
barplot2
}
\author{
Jiahao Wang
}
