% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dotplot2}
\alias{dotplot2}
\title{dotplot2}
\usage{
dotplot2(
  data,
  w = 6,
  h = 6,
  n = 20,
  colorHigh = "#DD1C77",
  colorLow = "#3182bd",
  sortBy = "p.adjust",
  title = "",
  padjust = TRUE,
  outName = NULL,
  PDF = FALSE,
  PNG = TRUE
)
}
\description{
dotplot2
}
\author{
Jiahao Wang
}
