% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dotplotGO}
\alias{dotplotGO}
\title{dotplotGO}
\usage{
dotplotGO(
  data,
  size = 28,
  padj = TRUE,
  outName = NULL,
  w = NULL,
  h = 8,
  PDF = FALSE,
  PNG = TRUE
)
}
\description{
dotplotGO
}
\author{
Jiahao Wang
}
