% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R
\name{FindClusters2}
\alias{FindClusters2}
\title{FindClusters2}
\usage{
FindClusters2(
  obj,
  cluster.range = c(35, 40),
  by = 0.1,
  res = 1,
  verbose = FALSE
)
}
\description{
FindClusters2
}
\author{
Jiahao Wang
}
