% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{loopControlor}
\alias{loopControlor}
\title{loopControlor}
\usage{
loopControlor(
  Ns,
  i,
  tag,
  Nlimit = 480,
  sleep = 60,
  flag = "amd_512",
  quota = 2,
  quiet = FALSE
)
}
\description{
loopControlor
}
\author{
Jiahao Wang
}
