% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R
\name{mergeSingleCellranger}
\alias{mergeSingleCellranger}
\title{mergeSingleCellranger}
\usage{
mergeSingleCellranger(dir, min.cells = 1, min.features = 1, method = "count")
}
\description{
mergeSingleCellranger
}
\author{
Jiahao Wang
}
