% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotVolcano}
\alias{plotVolcano}
\title{plotVolcano}
\usage{
plotVolcano(
  diffM,
  DIFFcutoff = 2,
  FDRcutoff = 0.05,
  title = "DGE_volcano",
  xtitle = "log2(FoldChange)",
  xlim = 5,
  ylim = 10,
  padjust = FALSE,
  size = 0.1,
  pos = 8.5,
  label = TRUE
)
}
\description{
plotVolcano
}
\author{
Jiahao Wang
}
