% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R
\name{QC_workflow}
\alias{QC_workflow}
\title{QC_workflow}
\usage{
QC_workflow(
  object,
  mt_n = 30,
  nFnC_n = 0.8,
  hyl = 200,
  hyh = 7000,
  vx = 40000,
  method = "check"
)
}
\description{
QC_workflow
}
\author{
Jiahao Wang
}
