% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{read.faster}
\alias{read.faster}
\title{read.faster}
\usage{
read.faster(
  file,
  header = FALSE,
  sep = "\\t",
  showProgress = FALSE,
  skip = 0,
  nrows = -1
)
}
\description{
read.faster
}
\author{
Jiahao Wang
}
