% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{region_plot_chromosome_chordal}
\alias{region_plot_chromosome_chordal}
\title{region_plot_chromosome_chordal}
\usage{
region_plot_chromosome_chordal(bed, PDF = NULL, Rplot = TRUE)
}
\description{
region_plot_chromosome_chordal
}
\author{
Jiahao Wang
}
