% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{region_plot_length_bar}
\alias{region_plot_length_bar}
\title{region_plot_length_bar}
\usage{
region_plot_length_bar(bed, PDF = NULL, Rplot = TRUE)
}
\description{
region_plot_length_bar
}
\author{
Jiahao Wang
}
