% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R
\name{RemoveDoublets}
\alias{RemoveDoublets}
\title{RemoveDoublets}
\usage{
RemoveDoublets(
  obj,
  doublet.rate,
  sample_name,
  pN = 0.25,
  pc.num = 1:30,
  use.SCT = FALSE,
  remove = TRUE
)
}
\description{
RemoveDoublets
}
\author{
Jiahao Wang
}
