% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{runGSEA}
\alias{runGSEA}
\title{runGSEA}
\usage{
runGSEA(
  DEGtable = NULL,
  geneSet = NULL,
  rnkFile = NULL,
  chipFile = NULL,
  gmtFile = NULL,
  prefix = "GSEA",
  outFolder = "GSEA",
  min = 15,
  max = 500,
  verbo = FALSE,
  PDF = FALSE
)
}
\description{
runGSEA
}
\author{
Jiahao Wang
}
