% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R
\name{sc_VolcanoPlot}
\alias{sc_VolcanoPlot}
\title{sc_VolcanoPlot}
\usage{
sc_VolcanoPlot(
  diff,
  log2FC = log2(1.5),
  padj = 0.05,
  label.symbols = NULL,
  label.max = 30,
  cols = c("#F93463", "#27A2EF"),
  title = ""
)
}
\description{
sc_VolcanoPlot
}
\author{
Jiahao Wang
}
