% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{setBlank}
\alias{setBlank}
\title{setBlank}
\usage{
setBlank(
  x.title = 1,
  x.text = 1,
  x.ticks = 1,
  y.title = 1,
  y.text = 1,
  y.ticks = 1
)
}
\description{
setBlank
}
\author{
Jiahao Wang
}
