% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{setLegend}
\alias{setLegend}
\title{setLegend}
\usage{
setLegend(
  direction = "vertical",
  position = "right",
  title = NA,
  order = NA,
  labels = NA,
  text.size = 10,
  text.col = "black",
  text.face = "plain",
  text.angle = 0,
  text.hjust = 0,
  text.vjust = 0,
  title.size = 10,
  title.col = "black",
  title.face = "plain",
  title.angle = 0,
  title.hjust = 0,
  title.vjust = 0,
  key.size = 2,
  key.color = NA,
  key.fill = NA,
  margin = 10
)
}
\description{
setLegend
}
\author{
Jiahao Wang
}
