% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{setText}
\alias{setText}
\title{setText}
\usage{
setText(
  title.size = 10,
  title.face = "bold",
  title.hjust = 0.5,
  title.vjust = 0,
  x.title.size = NA,
  x.title.face = NA,
  x.title.vjust = 0.5,
  x.text.size = NA,
  x.text.angle = 0,
  x.text.hjust = 0.5,
  x.text.vjust = 1,
  y.title.size = NA,
  y.title.vjust = 0,
  y.text.size = NA,
  face = "plain",
  y.title.face = NA,
  y.text.angle = 0,
  y.text.hjust = 1,
  graph.theme = "classic"
)
}
\description{
setText
}
\author{
Jiahao Wang
}
