% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{submitCMD}
\alias{submitCMD}
\title{submitCMD}
\usage{
submitCMD(
  cmd,
  jobName,
  c = 1,
  n = 1,
  N = 1,
  exclusive = NULL,
  verbose = FALSE,
  check = TRUE
)
}
\description{
submitCMD
}
\author{
Jiahao Wang
}
