% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R
\name{test_workflow}
\alias{test_workflow}
\title{test_workflow}
\usage{
test_workflow(
  object = NULL,
  nPC.method = "sigPC",
  N = 40,
  markers.1st2check = general_marker,
  assay = "RNA"
)
}
\description{
test_workflow
}
\author{
Jiahao Wang
}
