% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{venn}
\alias{venn}
\title{venn}
\usage{
venn(
  ...,
  Rplot = TRUE,
  PDF = NULL,
  PPT = NULL,
  cex.label = 8,
  cex.num = 8,
  alpha.fill = 0.7,
  col.fill = c("#E5D2DD", "#53A85F", "#F1BB72", "#F3B1A0"),
  alpha.stroke = 1,
  size.stroke = 1,
  col.stroke = "black",
  show_percentage = FALSE,
  returnP = FALSE
)
}
\description{
venn
}
\author{
Jiahao Wang
}
