% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{writeXlsx}
\alias{writeXlsx}
\title{writeXlsx}
\usage{
writeXlsx(
  ...,
  dfList = NULL,
  sheetNames = NULL,
  file = NULL,
  saveRowCol = NULL,
  append = FALSE
)
}
\description{
writeXlsx
}
\author{
Jiahao Wang
}
